backend/requirements.txt
annotated-types==0.7.0
anthropic==0.72.1
anyio==4.11.0
bcrypt==4.1.1
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
email-validator==2.1.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
gotrue==2.12.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
multidict==6.7.0
packaging==25.0
postgrest==2.24.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==2.0.7
realtime==2.24.0
six==1.17.0
sniffio==1.3.1
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.22.0
google-generativeai
